requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
matplotlib~=3.8.2
multidict==6.0.4
fastapi==0.108.0
pydantic==2.5.3
pydantic_core==2.14.6
pydantic_settings==2.1.0
typing_extensions==4.9.0
yarl==1.9.4
requests~=2.31.0
httpx~=0.26.0

pytest~=8.0.0
python-dotenv~=1.0.1